refactor test classes
