+++ src/test/scala/lilo/LiloApiSpec.scala
@@ -1,6 +1,5 @@
 package lilo
 
-import org.specs2.mutable.Specification
 import org.specs2.mock.Mockito
 import org.specs2.specification.Scope
 import com.twitter.util.Await
@@ -10,10 +9,7 @@ import com.twitter.util.Future
 import org.mockito.Mockito._
 
 @RunWith(classOf[JUnitRunner])
-class LiloSpec extends Specification with Mockito {
-
-  def resultOf[T](lilo: Lilo[T]) =
-    Await.result(lilo.run)
+class LiloApiSpec extends Spec {
 
   "the Lilo object" >> {
 
@@ -99,55 +95,6 @@ class LiloSpec extends Specification with Mockito {
     "can have its result filtered (lilo.withFilter)" in {
       resultOf(Lilo.value(1).withFilter(_ != 1)) mustEqual None
       resultOf(Lilo.value(1).withFilter(_ == 1)) mustEqual Some(1)
-    }
-  }
-
-  "The lilo execution model" >> {
-
-    "batches requests" >> {
-
-      "for multiple lilos created from traversed inputs" in {
-        ok
-      }
-
-      "for multiple lilos collected into only one lilo" in {
-        ok
-      }
-
-      "for lilos created inside nested flatmaps" in {
-        ok
-      }
-
-      "for lilos composed using for comprehension" >> {
-
-        "one level" in {
-          ok
-        }
-
-        "two levels" in {
-          ok
-        }
-
-        "many levels" in {
-          ok
-        }
-
-        "with a filter condition" in {
-          ok
-        }
-
-        "composition using a join" in {
-          ok
-        }
-      }
-    }
-
-    "executes joined lilos in parallel" in {
-      ok
-    }
-
-    "short-circuits the computation in case of a failure" in {
-      ok
     }
   }
 }
