+++ src/test/scala/lilo/LiloSourceSpec.scala
@@ -3,8 +3,6 @@ package lilo
 import org.specs2.mutable.Specification
 import org.specs2.mock.Mockito
 
-class LiloSourceSpec extends Specification with Mockito {
-  
-  
+class LiloSourceSpec extends Spec {
 
 }
